Ask before updating files

+++ src/hooks/init/check-local-sfdx-hardis-files.ts
@@ -1,5 +1,6 @@
 import * as c from 'chalk';
 import * as fs from 'fs-extra';
+import * as prompts from 'prompts';
 
 export const hook = async (options: any) => {
     // Skip hooks from other commands than hardis:scratch commands
@@ -24,7 +25,12 @@ async function managePackageJson(commandId: string) {
             const packageJson = JSON.parse(text);
             const hardisPackageJsonContent = await getSfdxHardisPackageJsonContent();
             packageJson['scripts'] = Object.assign(hardisPackageJsonContent['scripts'], packageJson['scripts']);
-            if (JSON.stringify(packageJson) !== JSON.stringify(JSON.parse(text))) {
+            if (JSON.stringify(packageJson) !== JSON.stringify(JSON.parse(text)) &&
+                !process.env.CI && (await prompts({
+                    type: 'confirm',
+                    name: 'update',
+                    message: 'Your package.json is deprecated, do you agree to upgrade it ? (If you hesitate, just trust us and accept)'
+                })).value === true) {
                 await fs.writeFile(packageJsonFile, JSON.stringify(packageJson, null, 2));
                 console.log(c.cyan('[sfdx-hardis] Updated package.json with sfdx-hardis content'));
             }
@@ -53,7 +59,11 @@ async function manageGitIgnore(commandId: string) {
                 updated = true;
             }
         }
-        if (updated) {
+        if (updated && !process.env.CI && (await prompts({
+            type: 'confirm',
+            name: 'update',
+            message: 'Your .gitignore is deprecated, do you agree to upgrade it ? (If you hesitate, just trust us and accept)'
+        })).value === true) {
             await fs.writeFile(gitIgnoreFile, gitIgnoreLines.join('\n') + '\n', 'utf-8');
             console.log(c.cyan('[sfdx-hardis] Updated .gitignore'));
         }
@@ -64,12 +74,14 @@ async function manageGitIgnore(commandId: string) {
 async function getSfdxHardisPackageJsonContent() {
     const hardisPackageJsonContent = {
         scripts: {
-            'scratch:create': 'sfdx hardis:scratch:create',
             'scratch:push': 'sfdx force:source:push -g -w 60 --forceoverwrite',
             'scratch:pull': 'sfdx force:source:pull --forceoverwrite',
             'scratch:open': 'sfdx force:org:open',
             'org:test:apex': 'sfdx hardis:org:test:apex',
-            'login:reset': 'sfdx auth:logout --noprompt || true && sfdx config:unset defaultusername defaultdevhubusername -g && sfdx config:unset defaultusername defaultdevhubusername || true'
+            'scratch:create': 'sfdx hardis:scratch:create',
+            'login:reset': 'sfdx auth:logout --noprompt || true && sfdx config:unset defaultusername defaultdevhubusername -g && sfdx config:unset defaultusername defaultdevhubusername || true',
+            'configure:org:devhub': 'sfdx hardis:project:configure:deployments --devhub',
+            'configure:org:deployment': 'sfdx hardis:project:configure:deployments'
         }
     };
     return hardisPackageJsonContent;
